Refactorization of the code, and Makefile added for compilation
