+++ dijkstra/include/params.h
@@ -1,0 +1,7 @@
+#ifndef PARAMS_H
+#define PARAMS_H
+
+#define V 6
+#define MAX_INT 1000
+
+#endif
