+++ dijkstra/src/utils.c
@@ -1,12 +1,4 @@
-#include <stdio.h>
-#define V 6
-#define MAX_INT 1000
-
-struct Graph
-{
-    int vertices;
-    int matrix[V][V];
-};
+#include "../include/utils.h"
 
 void add_vertex(struct Graph* graph, 
     int source, 
@@ -96,20 +88,4 @@ void dijkstraAlgorithm(
     }
 
     print_dijkstra(dist);
-}
-
-int main()
-{
-    struct Graph G;
-    initialize_graph(&G, V);
-    add_vertex(&G, 0, 1, 1);
-    add_vertex(&G, 1, 3, 1);
-    add_vertex(&G, 0, 2, 5);
-    add_vertex(&G, 2, 3, 1);
-    add_vertex(&G, 0, 4, 3);
-    add_vertex(&G, 4, 5, 1);
-    add_vertex(&G, 2, 5, 2);
-    
-    dijkstra(&G, 0);
-    return 0;
 }
